ModelCreatationTab refactor
