+++ Resources/modules/ModelCreationTab/FeatureSelectionAlgorithms/FeatSel_SequentialForwardFloating.py
@@ -1,29 +1,29 @@
 """
-Script Name:    FeatSel_SequentialBackwardFloating.py
+Script Name:    FeatSel_SequentialForwardFloating.py
 
-Description:    Sequential Backward Floating Selection is a feature
-                selection scheme that iteratively removes features 
-                to an initially full model. Features are removed if
-                thier removal to the model increases the model's
-                score. Features are then added iteratively and
+Description:    Sequential Forward Floating Selection is a feature
+                selection scheme that iteratively adds features 
+                to an initially empty model. Features are added if
+                thier addition to the model increases the model's
+                score. Features are then removed iteratively and
                 the best resulting model is kept. A rough algorith 
                 for this scheme is as follows 
                 (taken from: http://research.cs.tamu.edu/prism/lectures/pr/pr_l11.pdf):
 
                 For a system with 6 possible predictors
 
-                1) Y = "111111" [no predictors in model]
+                1) Y = "000000" [no predictors in model]
 
-                2) Select the best feature to remove from the model:
-                    predictor_to_remove = maximize_score(Y - k)
-                        (where k is one of the 6 predictors)
-                    Y = Y - predictor_to_remove 
-
-                3) Select the best feature to add:
+                2) Select the best feature to add to the model:
                     predictor_to_add = maximize_score(Y + k)
+                        (where k is one of the 6 predictors)
+                    Y = Y + predictor_to_add 
 
-                4) if score(Y+k) > score(Y) then 
-                    Y = Y + predictor_to_add
+                3) Select the worst feature to remove:
+                    predictor_to_remove = maximize_score(Y - k)
+
+                4) if score(Y-k) > score(Y) then 
+                    Y = Y - predictor_to_remove
                     goto 3
                    else
                     goto 2
@@ -31,20 +31,17 @@ Description:    Sequential Backward Floating Selection is a feature
 
 import bitarray as ba
 import importlib
-from resources.modules.StatisticalModelsTab import ModelScoring
+from resources.modules.ModelCreationTab import ModelScoring
 import numpy as np
+from itertools import compress
 
 
 class FeatureSelector(object):
 
-    NAME = "Sequential Backward Floating Selection"
-    DESCRIPTION = "Starts with a full set of predictors and iteratively removes low-performing predictors until the best model is found."
+    NAME = "Sequential Forward Floating Selection"
+    DESCRIPTION = "Starts with no predictors and iteratively adds in the best performing predictors until no more performance gains can be made."
 
     def __init__(self, parent = None, **kwargs):
-        """
-        """
-
-        #
 
         # Create References to the predictors and the target
         self.predictorPool = parent.modelRunTableEntry['PredictorPool']
@@ -60,7 +57,7 @@ class FeatureSelector(object):
         
         # Create variables to store the current predictors and performance
         self.numPredictors = len(self.predictorPool)
-        self.currentPredictors = kwargs.get("initialModel", ba.bitarray([True] * self.numPredictors))
+        self.currentPredictors = kwargs.get("initialModel", ba.bitarray([False] * self.numPredictors))
 
         # Add in any forced predictors
         self.currentPredictors = self.currentPredictors | self.parent.forcedPredictors
@@ -68,7 +65,6 @@ class FeatureSelector(object):
         # Keep track of model performance with the previos predictors and the current model scores
         self.previousPredictors = self.currentPredictors.copy()
         self.currentScores = {}
-
 
         return
     
@@ -89,6 +85,10 @@ class FeatureSelector(object):
         else:
             y = self.parent.proc_yTraining[~np.isnan(x).any(axis=1)]
             x = x[~np.isnan(x).any(axis=1)]
+
+        # If any of the variables contain only zero's, we get singular matrix issues (obviosuly).
+        if any(np.all((x==0), axis=0)):
+            return {self.regression.scoringParameters[i]: np.nan for i, scorer in enumerate(self.regression.scorers)}
 
         # Fit the model with the regression method and get the resulting score
         try:
@@ -120,7 +120,7 @@ class FeatureSelector(object):
 
         # Store the results in the more comprehensive resultsList
         self.parent.resultsList.append(
-            {"Model":modelStr, "Score":score, 
+            {"Model":list(model), "Score":score, 
              "Method":"PIPE/{0}/{1}/{2}".format(self.parent.preprocessor.FILE_NAME, 
                                       self.regressionName,  
                                       self.regression.crossValidation)})
@@ -145,10 +145,10 @@ class FeatureSelector(object):
         while True:
             
             # Search for new predictors to add
-            self.trySubtraction()
+            self.tryAddition()
 
             # Search for predictors to remove
-            self.tryAddition()
+            self.trySubtraction()
 
             # Check if we've added any new predictors or removed predictors
             if self.previousPredictors == self.currentPredictors:
@@ -176,14 +176,11 @@ class FeatureSelector(object):
         2) We can potentially add the 1st, 2nd and 4th predictors
         3) compute the scores of adding each predictor individaully
         4) If any of the scores is greater than the initial score,
-           that predictor is added and the function tries to remove another one.
+           that predictor is added and the function returns.
         """
 
         # Make a copy of the predictors that we can manipulate
         model = self.currentPredictors.copy()
-
-        # Keep track of whether we added a predictor
-        predictorAdded = False
 
         # Iterate over the predictors
         for i in range(self.numPredictors):
@@ -193,13 +190,13 @@ class FeatureSelector(object):
 
                 # Add the predictor to the model
                 model[i] = True
-                modelStr = model.to01()
+                model_str = model.to01()
 
                 # Check that we haven't already computed this model combination
-                if modelStr in self.parent.computedModels:
+                if model_str in self.parent.computedModels:
                 
                     # Get the score from the list of models
-                    score = self.parent.computedModels[modelStr]
+                    score = self.parent.computedModels[model_str]
                     
                 else:
                 
@@ -213,19 +210,11 @@ class FeatureSelector(object):
                 if ModelScoring.scoreCompare(newScores = score, oldScores = self.currentScores):
                     self.currentScores = score
                     self.currentPredictors = model.copy()
-                    predictorAdded = True
 
                 # Revert the model
                 model[i] = False
-
-        # If we added a predictor, try to add another one
-        if predictorAdded:
-            self.tryAddition()
-
-        # Otherwise, return
-        else:
-            return
         
+        return
     
 
     def trySubtraction(self):
@@ -241,11 +230,14 @@ class FeatureSelector(object):
         2) We can potentially remove the 3rd, 4th and 5th predictors
         3) compute the scores of removing each predictor individaully
         4) If any of the scores is greater than the initial score,
-           that predictor is removed
+           that predictor is removed and the function trys to remove another one.
         """
 
         # Make a copy of the predictors that we can manipulate
         model = self.currentPredictors.copy()
+
+        # Keep track of whether we removed a predictor
+        predictorRemoved = False
 
         # Iterate over the predictors
         for i in range(self.numPredictors):
@@ -255,13 +247,13 @@ class FeatureSelector(object):
                 
                 # Remove the predictor from the model
                 model[i] = False
-                modelStr = model.to01()
+                model_str = model.to01()
 
                  # Check that we haven't already computed this model combination
-                if modelStr in self.parent.computedModels:
+                if model_str in self.parent.computedModels:
                 
                     # Get the score from the list of models
-                    score = self.parent.computedModels[modelStr]
+                    score = self.parent.computedModels[model_str]
                     
                 else:
                 
@@ -275,10 +267,17 @@ class FeatureSelector(object):
                 if ModelScoring.scoreCompare(newScores = score, oldScores = self.currentScores):
                     self.currentScores = score
                     self.currentPredictors = model.copy()
+                    predictorRemoved = True
 
                 # Revert the model
                 model[i] = True
         
-        return
+        # If we removed a predictor, try to remove another one
+        if predictorRemoved:
+            self.trySubtraction()
+
+        # Otherwise, return
+        else:
+            return
